Conflicts:
	inc/U_USART1.h
	src/U_USART1.cpp

+++ inc/Delay.h
@@ -11,13 +11,11 @@
 #include "stm32f10x.h"
 
 extern void Delay_Init(uint16_t DelayResolution_us);
-extern void Delay_us(uint32_t us);
-extern inline void Delay_ms(uint32_t ms)
-{
+extern void Delay_us(uint64_t us);
+extern inline void Delay_ms(uint32_t ms) {
 	Delay_us(ms * 1000);
 }
-extern inline void Delay_s(uint32_t s)
-{
+extern inline void Delay_s(uint32_t s) {
 	Delay_us(s * 1000000);
 }
 extern uint32_t millis();
